requirements.txt
# Python requirements for Video Conversation Agent

# Core dependencies
python-dotenv==1.0.0
google-generativeai==0.3.0

# YouTube transcript extraction
youtube-transcript-api==0.6.1
pytube==15.0.0

# Logging and utilities
requests==2.31.0

# Optional: For enhanced CLI experience
colorama==0.4.6
